app/requirements.txt
Flask~=1.1.2
psycopg2-binary~=2.8.6
sqlalchemy~=1.3.23
discogs-client~=2.3.0
pyOpenSSL~=22.0.0
python-telegram-bot~=13.11
